Design main page, home and update features
projects still in the works: making each tweet have a responsive
user link that will take you to a page of only their tweets, which
will then update with the update button.

After that, comes the task of adding in a settings page, that will
allow a user to set their id, max tweets per page, maybe even color
schemes.